RF: add thin layer around xml.etree.ElementTree and share XmlSerializable
